[ENH]: Wrap spanner Row to a new type
