+++ rust/worker/Cargo.toml
@@ -61,7 +61,7 @@ chroma-segment = { workspace = true }
 chroma-storage = { workspace = true }
 chroma-system = { workspace = true }
 chroma-sysdb = { workspace = true }
-chroma-types = { workspace = true, features = ["spanner"] }
+chroma-types = { workspace = true }
 chroma-jemalloc-pprof-server = { workspace = true }
 s3heap = { workspace = true }
 s3heap-service = { workspace = true }
